Proposed restructure for the Package - breaking changes
