+++ Examples/Demo/Demo/LazyLoadingView.swift
@@ -1,3 +1,4 @@
+import ImageProviders
 import MarkdownUI
 import SwiftUI
