+++ Sources/ImageProviders/AssetInlineImageProvider.swift
