+++ Sources/ImageProviders/ImageProvider.swift
@@ -26,16 +26,16 @@ public protocol ImageProvider {
   @ViewBuilder func makeImage(url: URL?) -> Body
 }
 
-struct AnyImageProvider: ImageProvider {
+public struct AnyImageProvider: ImageProvider {
   private let _makeImage: (URL?) -> AnyView
 
-  init<I: ImageProvider>(_ imageProvider: I) {
+  public init<I: ImageProvider>(_ imageProvider: I) {
     self._makeImage = {
       AnyView(imageProvider.makeImage(url: $0))
     }
   }
 
-  func makeImage(url: URL?) -> some View {
+  public func makeImage(url: URL?) -> some View {
     self._makeImage(url)
   }
-}
+}
